AÑADIDO ICONO DE INTERACCIÓN -añadido icono en la UI para indicar si el objeto frente al usuario es interactivo o no. -arreglos varios

+++ Assets/Scripts/Systems/Dialogue.cs
@@ -8,7 +8,7 @@ public class Dialogue : MonoBehaviour
     [Header("Properties")]
     //
     [SerializeField] private TMP_Text messageDisplay;
-    [SerializeField, Min(1)] private int charactersPerSecond;
+    [SerializeField, Min(1)] private float charactersPerSecond;
     [SerializeField] private float textDurationAfterTyping;
     [SerializeField, Range(float.Epsilon, 10)] private float fadeOutSpeed;
     private const float FADE_SPEED_MULTIPLIER = 0.01f;
